XWIKI-17052: Allow to inject Javascript / CSS in Notification Displayer
The idea of this refactoring was first to introduce a new component to
allow using Async Rendering framework for Notifications. But in order to
have a common way of requesting notifications either from the REST
service or from the Async renderer, I decided to put the
NotificationParameters as the central parameter object for both
services, I created a dedicated factory for them and I externalized the
Cache management of notifications which now also use the
NotificationParameters.

  * Refactoring of notification-rest-module to simplify it and to
    delegate to the right entities
  * Creation of DefaultNotificationParametersFactory component to create
    NotificationParameters
  * Creation of DefaultNotificationCacheManager component to handle
    notifications caches
  * Creation of InternalHtmlNotificationRenderer component to handle
    rendering of CompositeEvent in HTML only
  * Introduction of DefaultAsyncNotificationRenderer and related classes
    to use the async rendering framework for Notifications
  * Add of new methods in script services to be able to use the async
    framework for notifications and to deprecate the usage of REST
service for our notifications
  * Add of hashCode and Equals in various filters and notification
    preferences classes and in NotificationParameters so we can rely on
them for the cache key